requirements_old.txt
asn1crypto==0.24.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
contextlib2==21.6.0
cryptography==2.7
Django==2.2
django-appconf==1.0.3
django-compressor==2.3
django-constance==2.1.0
django-crispy-forms==1.7.0
django-easy-select2==1.3.4
django-extensions
django-guardian==1.5.1
django-imagekit==3.3
django-inline-static==0.1.0
django-jsonview==2.0.0
django-kronos==1.0
django-localflavor==1.2
django-markdown2==0.3.0
django-qr-code==0.3.3
django-sass==1.0.1
django-simple-history==1.9.0
django-widget-tweaks==1.4.1
django-wkhtmltopdf==3.1.0
docutils==0.14
et-xmlfile==1.0.1
idna==2.8
jdcal==1.4.1
LEPL==5.1.3
libsass==0.21.0
lxml==4.6.3
markdown2==2.3.6
model-mommy==1.2.6
olefile==0.46
openpyxl==2.6.2
packaging==19.0
petl==1.1.1
pilkit==2.0
Pillow==5.4.1
psycopg2==2.8
pyasn1==0.4.5
pycparser==2.19
Pygments==2.4.2
pyOpenSSL==19.0.0
pyparsing==2.3.1
PySocks==1.7.0
python-crontab==2.1.1
python-dateutil==2.8.0
python-magic==0.4.15
pytz==2019.1
qrcode==6.1
raven==5.26.0
rcssmin==1.0.6
redgreenunittest==0.1.1
redis==3.3.4
requests==2.22.0
rjsmin==1.1.0
sentry-sdk==0.8.0
setuptools-scm==3.3.3
six==1.12.0
Unidecode==1.0.23
urllib3==1.24.3
waitress==1.2.1
XlsxWriter==0.8.9

django-extra-views
django-import-export
django-nested-inline
django-parsley
django-report-builder
django-summernote
djangocms-admin-style
pandas
xlwt
